Reorganize code structures
- Add test utility project
