+++ dictzip-cli/src/main/java/org/dict/zip/cli/DictZipUtils.java
@@ -32,7 +32,7 @@ public final class DictZipUtils {
      * @param name input filename.
      * @return output filename.
      */
-    protected static String uncompressedFileName(final String name) {
+    static String uncompressedFileName(final String name) {
         String result;
         if (name.endsWith(".dz") || name.endsWith(".gz")) {
             result = name.substring(0, name.length() - 3);
@@ -47,7 +47,7 @@ public final class DictZipUtils {
      * @param name input file name.
      * @return output filename.
      */
-    protected static String compressedFileName(final String name) {
+    static String compressedFileName(final String name) {
         return name + ".dz";
     }
