+++ gradle.properties
@@ -7,5 +7,5 @@ projectDesc=DictZip compression library and CLI
 projectOwner=miurahr
 developerEmail=miurahr@linux.com
 developerName='Hiroshi Miura'
-projectYears=2016-2021
+projectYears=2016-2022
 projectReleaseBody="Fix release for critical bug on seek."
