+++ northside-io/src/main/java/tokyo/northside/io/IOUtils2.java
